duplicate fixture
