+++ tardis/visualization/tools/tests/test_liv_plot.py
@@ -465,23 +465,8 @@ class TestLIVPlotter:
         expected.close()
 
     @pytest.fixture(scope="class")
-    def workflow_simple(self, config_verysimple, atomic_data_fname):
-        config = deepcopy(config_verysimple)
-        config.atom_data = atomic_data_fname
-        config.montecarlo.iterations = 3
-        config.montecarlo.no_of_packets = 4000
-        config.montecarlo.last_no_of_packets = -1
-        config.spectrum.virtual.virtual_packet_logging = True
-        config.montecarlo.no_of_virtual_packets = 1
-        config.spectrum.num = 2000
-        
-        workflow = StandardTARDISWorkflow(config, enable_virtual_packet_logging=True)
-        workflow.run()
-        return workflow
-
-    @pytest.fixture(scope="class")
-    def plotter_from_workflow(self, simulation_simple):
-        return LIVPlotter.from_simulation(simulation_simple)
+    def plotter_from_workflow(self, workflow_simple):
+        return LIVPlotter.from_simulation(workflow_simple)
 
     @pytest.fixture(scope="class")
     def liv_regression_data(self, tardis_regression_path):
